sim/pyproject.toml
[project]
name = "darwin-sim"
version = "0.4.0"
description = "DARWIN evolutionary DEX simulator — proves or falsifies the market-evolution thesis before production."
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
    "zstandard>=0.22",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
plots = ["matplotlib>=3.9"]

[project.scripts]
darwin-sim = "darwin_sim.experiments.runner:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"
